rag-customers/rag-customers/requirements.txt
FastAPI
uvicorn
openai
supabase-py
pydantic
python-dotenv